+++ packages/multipart-parser/bench/parsers/multipart-parser.ts
@@ -5,9 +5,9 @@ import { MultipartMessage } from '../messages.ts';
 export async function parse(message: MultipartMessage): Promise<number> {
   let start = performance.now();
 
-  await parseMultipart(message.generateChunks(), { boundary: message.boundary }, (_part) => {
-    // Do nothing
-  });
+  for (let _ of parseMultipart(message.generateChunks(), { boundary: message.boundary })) {
+    // Do nothing with the part, just iterate through it to measure parsing time
+  }
 
   return performance.now() - start;
 }