+++ packages/multipart-parser/examples/cf-workers/src/index.ts
@@ -31,24 +31,13 @@ export default {
         let bucket = env.MULTIPART_UPLOADS;
         let parts: any[] = [];
 
-        await parseMultipartRequest(request, async (part) => {
+        for await (let part of parseMultipartRequest(request)) {
           if (part.isFile) {
             let uniqueKey = `upload-${new Date().getTime()}-${Math.random()
               .toString(36)
               .slice(2, 8)}`;
 
-            // Put the file in R2.
-
-            // Ideally we could stream part.body directly, but Cloudflare's R2
-            // API requires a FixedLengthStream and unfortunately we don't know
-            // the length of the stream at this point because browsers don't send
-            // Content-Length headers with file uploads.
-            // await bucket.put(uniqueKey, part.body);
-
-            // So instead, we have to buffer the entire file in memory and then
-            // upload it to R2.
-            let bytes = await part.bytes();
-            await bucket.put(uniqueKey, bytes, {
+            await bucket.put(uniqueKey, part.bytes, {
               httpMetadata: {
                 contentType: part.headers.get('Content-Type')!,
               },
@@ -58,15 +47,12 @@ export default {
               name: part.name,
               filename: part.filename,
               mediaType: part.mediaType,
-              size: bytes.byteLength,
+              size: part.size,
             });
           } else {
-            parts.push({
-              name: part.name,
-              value: await part.text(),
-            });
+            parts.push({ name: part.name, value: part.text });
           }
-        });
+        }
 
         return new Response(JSON.stringify({ parts }, null, 2), {
           headers: { 'Content-Type': 'application/json' },